requirements.txt
# Noise generation & text augmentation
nlpaug>=1.1.11
textattack>=0.3.9

# NLP processing
spacy>=3.7.0
nltk>=3.8.1

# Grammar scoring
language-tool-python>=2.7.1

# Semantic similarity
sentence-transformers>=2.2.2

# Fluency / readability
textstat>=0.7.3

# Math / statistics
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# LLM API clients
openai>=1.0.0
google-generativeai>=0.4.0

# Wikipedia corpus
wikipedia-api>=0.6.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
